SSL: kTLS: fix possible close() lock
With new kernel TLS option enable it is became possible to lock on
close() syscal and wait until kernel tries to send all buffered data
on slow or even dead socket. This makes stalls on workers and unability
to accept new connection. In test cases workers were stalled for
minutes. The only possible solution to prevent kernel waiting on such
sockets is to configure SO_SNDTIMEO sockopt with a small timeout just
before closing socket.

This was possibly fixed in kernels 5.2+ but did not test it.

Signed-off-by: Vadim Fedorenko <junjunk@github.com>