Core: log amount of ready to accept connections
The side-effect of SO_REUSEPORT option is that connections in
accept queue could not be spread to other waiting socket in group
when one socket is closing. All such already established connections
will be dropped. This occures on server reload and is not accouted
anywhere so let's log it just before closing waiting socket.